[INTERNAL] Split code of event model into several files
